takoapp1/requirements.txt
django
djangorestframework
openai
python-dotenv
django-cors-headers
channels
channels-redis
pillow
